World chunk protocol update.
PlayerConnect area modification. Doesn't spam error now..

A player already spawned should not trigger PlayerJoin

Using Constants
