+++ src/main/java/org/spout/vanilla/protocol/bootstrap/handler/BootstrapLoginRequestMessageHandler.java
@@ -26,17 +26,10 @@
  */
 package org.spout.vanilla.protocol.bootstrap.handler;
 
-import org.spout.api.Spout;
-import org.spout.api.event.Event;
-import org.spout.api.event.player.PlayerConnectEvent;
 import org.spout.api.player.Player;
 import org.spout.api.protocol.MessageHandler;
 import org.spout.api.protocol.Session;
 
-import org.spout.vanilla.VanillaPlugin;
-import org.spout.vanilla.configuration.VanillaConfiguration;
-import org.spout.vanilla.protocol.VanillaProtocol;
-import org.spout.vanilla.protocol.bootstrap.handler.auth.LoginAuth;
 import org.spout.vanilla.protocol.msg.login.LoginRequestMessage;
 
 public class BootstrapLoginRequestMessageHandler extends MessageHandler<LoginRequestMessage> {
