+++ src/test/java/org/spout/vanilla/protocol/VanillaProtocolTest.java
@@ -157,7 +157,7 @@ public class VanillaProtocolTest extends BaseProtocolTest {
 			new EntityRemoveEffectMessage(1, (byte) 1),
 			new SetExperienceMessage(1.2F, (short) 2, (short) 3),
 			new LoadChunkMessage(0, -2, true),
-			new CompressedChunkMessage(1, 2, true, new boolean[16], 1, new byte[][]{new byte[16 * 16 * 16 * 5 / 2], null, null, null, null, null, null, null, null, null, new byte[Chunk.BLOCKS.HALF_VOLUME * 5], null, null, null, null, null}, new byte[16 * 16]),
+			new CompressedChunkMessage(1, 2, true, new boolean[16], new byte[][]{new byte[16 * 16 * 16 * 5 / 2], null, null, null, null, null, null, null, null, null, new byte[Chunk.BLOCKS.HALF_VOLUME * 5], null, null, null, null, null}, new byte[16 * 16]),
 			new MultiBlockChangeMessage(2, 3, new short[]{2, 3, 4, /**/ 3, 6, 4, /**/ 8, 5, 5}, new short[]{1, 2, 3}, new byte[]{3, 4, 5}),
 			new BlockChangeMessage(1, 2, 3, 87, 2),
 			new BlockActionMessage(1, 2, 3, (byte) 4, (byte) 5, (byte) 29),
